requirements.txt
# Core NLP & ML
spacy==3.7.2
spacy-transformers==1.3.4
tensorflow==2.15.0
scikit-learn==1.4.0
transformers==4.36.2

# OCR & PDF Processing
pytesseract==0.3.10
pdf2image==1.16.3
PyMuPDF==1.23.21
Pillow==10.2.0

# API & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3

# Development
jupyter==1.0.0
ipykernel==6.28.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Model Evaluation
seqeval==1.2.2
